Initial commit of graph_mod module
Coauthors: Donald Lafleur and Jonathan Llovet.
Calculates the number of times that a graph with a given number
of nodes is traversed by using a given step size between nodes.
This is done by using modular arithmetic, so that each node in
a graph is represented by an integer in base x, where x is the
number of nodes in the graph.

To do: Clean up print formatting; extract footer to function;
Introduce interface for arrays containing results;
add __main__ check