Create data structure to hold coordinates for hits around circle
The math module has been imported to calculate the sine and cosine
of each of the hits around a circle. These are calculated based on
the angle between each of the steps around the circle used as an
image of the modular arithmetic being examined.

hit_data contains a list of dictionaries that contain the angle and
the x and y coordinates of each of the steps around the graph.

Todo: Fix calculation errors
Rename graph_journey
Import matplotlib